# seaf-archtool-core/public/documentation/docs/manual/backend/api.yaml
...
openapi: 3.0.3
info:
  version: 1.0.0
  title: SEAF.ArchTool API
  description: >
    Данный документ актуален для развертывания в режимах Client-Server и Cluster.
    Здесь предоставляются интеграционные API, которые вы можете использовать для 
    встраивания SEAF.ArchTool в свои производственные процессы.
    Подробнее тут https://gitverse.ru/seafteam/seaf-archtool-core/content/master/public/documentation/docs/manual/config/deployment.md
servers:
  - url: https://{URL}
    description: API server secure
    variables:
      URL:
        default: localhost:3030
        description: Сервер https для запроса
  - url: http://{URL}
    description: API server insecure
    variables:
      URL:
        default: localhost:3030
        description: Сервер http для запроса
...
paths:
...
  /seaf-core/api/entities/{entity}/presentations/{presentation}:
    get:
      tags:
      - Entities
      summary: Вызывает специальную (upload) презентацию сущности 
      parameters:
        - in: path
          name: entity
          description: Идентификатор сущности
          required: true
          schema:
            $ref: '#/components/schemas/EntityID'
          example: interactions
        - in: path
          name: presentation
          description: Идентификатор презентации
          required: true
          schema:
            $ref: '#/components/schemas/PresentationID'
          example: export
        - in: query
          name: id
          required: true
          description: >
            Параметр передающийся в презентацию. Зависит от самой презентации.
            Будет доступен в презентации в $params.
          schema:
            $ref: '#/components/schemas/EntityParam'
          example: archtool.user
      responses:
        "200":
          description: "Презентация сущности"
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/ProblemsResult'
...
components:
  schemas:
...
    EntityID: 
      type: string
      pattern: '^[a-zA-Z][a-zA-Z\_0-9]*$'
    PresentationID: 
      type: string
      pattern: '^[a-zA-Z][a-zA-Z\_0-9]*$'
    EntityParam:
      type: string
    ProblemsResult:
      type: object
      properties:
        process:
          type: object
          properties:
            startEvent:
              type: object
              title: Просмотр архитектуры
              properties:
                outgoing:
                  type: string
              required:
                - outgoing
            task:
              type: array
              title: Вход на портал
              items:
                - type: object
                  properties:
                    incoming:
                      type: string
                    outgoing:
                      type: string
                  required:
                    - incoming
                    - outgoing
                - type: object
                  properties:
                    incoming:
                      type: string
                    outgoing:
                      type: string
                  required:
                    - incoming
                    - outgoing
            sequenceFlow:
              type: array
              items:
                - type: string
            exclusiveGateway:
              type: object
              title: Информация о существующей архитектуре
              properties:
                incoming:
                  type: string
              required:
                - incoming
          required:
            - startEvent
            - task
            - sequenceFlow
            - exclusiveGateway
      required:
        - process
...